feat: incorporate click
Addresses (but does not close) issue #9. Most day-to-day blog admin
functions can now be executed using the command line, rather than
needing to open up a python console.

This will enable using context managers for tasks that require the flask
app or the database, such that Flask-SQLAlchemy 3.0 can be used