add default configuration file
Added a default config file with comments in config/default.cfg. Added a
very basic check to blog_admin initialization that verifies that the
config file exists, and copies over the default if it doesn't.

Not a perfect fix, but addressed the issue I had with not even knowing
how to format a config file a couple of nights ago. Closes #15.

Also moved the `IMAGES_DIRECTORY` to a baked-in constant rather than
something that's configurable; it's baked-in to the existing posts, so
no reason to make it optional anymore.